fix(a11y): déplace les appels de SwitchPeriod dans les simulateurs lodeom et rgcp
